Dedicated CLASS arrangement for errors. Added walking conditions.

+++ Rest/AgentRest/AgentRest/Models/MissionModel.cs
@@ -13,8 +13,8 @@
         public int TargetId { get; set; }
         public AgentModel Agent { get; set; }
         public TargetModel Target { get; set; }
-        public float TimeLeft { get; set; }
-        public float TimeRight { get; set; }
+        public double TimeLeft { get; set; }
+        public double TimeRight { get; set; }
 
         public MissionStatus Status { get; set; }
     }
